requirements.txt
pandas
torch
transformers
Pillow
requests
beautifulsoup4